Seperated quest action form rendering config into dedicated hook
