+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/QuestActionSubForm/QuestActionSubForm.tsx
@@ -1,19 +1,9 @@
-import {
-  QuestParticipationLimit,
-  QuestParticipationPeriod,
-} from '@hicommonwealth/schemas';
-import { PRODUCTION_DOMAIN } from '@hicommonwealth/shared';
+import { QuestParticipationLimit } from '@hicommonwealth/schemas';
 import clsx from 'clsx';
-import { numberNonDecimalGTZeroValidationSchema } from 'helpers/formValidations/common';
 import { doesActionRewardShareForReferrer } from 'helpers/quest';
 import { splitCamelOrPascalCase } from 'helpers/string';
-import useRunOnceOnCondition from 'hooks/useRunOnceOnCondition';
-import React, { useEffect } from 'react';
-import { fetchCachedNodes } from 'state/api/nodes';
-import CWRepetitionCycleRadioButton, {
-  useCWRepetitionCycleRadioButton,
-} from 'views/components/component_kit/CWRepetitionCycleRadioButton';
-import { ValidationFnProps } from 'views/components/component_kit/CWRepetitionCycleRadioButton/types';
+import React from 'react';
+import CWRepetitionCycleRadioButton from 'views/components/component_kit/CWRepetitionCycleRadioButton';
 import { CWIconButton } from 'views/components/component_kit/cw_icon_button';
 import { CWText } from 'views/components/component_kit/cw_text';
 import { CWSelectList } from 'views/components/component_kit/new_designs/CWSelectList';
@@ -27,311 +17,26 @@ import {
   QuestActionContentIdScope,
   QuestActionSubFormProps,
 } from './types';
+import useQuestActionSubForm from './useQuestActionSubForm';
 
-// these restrictions are only on client side, update per future requirements
-const MAX_REPETITION_COUNTS = {
-  PER_DAY: 4,
-  PER_WEEK: 28,
-  PER_MONTH: 120,
-};
-
-const QuestActionSubForm = ({
-  isRemoveable = true,
-  onRemove,
-  errors,
-  defaultValues,
-  config,
-  onChange,
-  availableActions,
-  hiddenActions,
-  internalRefs,
-}: QuestActionSubFormProps) => {
-  const actionOptions = availableActions
-    .map((event) => ({
-      value: event as QuestAction,
-      label: splitCamelOrPascalCase(event),
-    }))
-    .filter(
-      (action) =>
-        !(hiddenActions || []).includes(action.value) &&
-        action.value !== 'UserMentioned',
-    );
-
-  const contentIdInputConfig = {
-    placeholders: {
-      sampleThreadLink: `https://${PRODUCTION_DOMAIN}/discussion/25730`,
-      sampleCommentLink: `https://${PRODUCTION_DOMAIN}/discussion/25730?comment=89775`,
-      sampleTopicLink: `https://${PRODUCTION_DOMAIN}/common/discussions/Proposals`,
-      twitterTweetUrl: `https://x.com/user/status/1904060455158428146`,
-      discordServerId: `0xxxxxxxxxxxxxxxx0`,
-      chainId: `Select community chain`,
-      groupId: `https://${PRODUCTION_DOMAIN}/common/members?tab=groups&groupId=1234`,
-    },
-    labels: {
-      threadId: 'Thread Link (optional)',
-      commentId: 'Comment Link (optional)',
-      topicId: 'Topic Link (optional)',
-      twitterTweetUrl: 'Tweet URL',
-      chainId: 'Chain Id (optional)',
-      discordServerId: 'Discord Server Id',
-      groupId: 'Group Link',
-    },
-  };
-
-  const startLinkInputConfig = {
-    placeholders: {
-      discordServerUrl: `https://discord.gg/commonwealth`,
-    },
-    labels: {
-      discordServerUrl: 'Discord Server Url',
-    },
-  };
-
-  const getContentIdInputLabel = () => {
-    if (defaultValues?.contentIdScope === QuestActionContentIdScope.Thread) {
-      if (config?.with_optional_thread_id) {
-        return contentIdInputConfig.labels.threadId;
-      }
-      if (config?.with_optional_comment_id) {
-        return contentIdInputConfig.labels.commentId;
-      }
-    }
-    if (
-      config?.with_optional_topic_id ||
-      defaultValues?.contentIdScope === QuestActionContentIdScope.Topic
-    ) {
-      return contentIdInputConfig.labels.topicId;
-    }
-
-    if (config?.requires_twitter_tweet_link) {
-      return contentIdInputConfig.labels.twitterTweetUrl;
-    }
-
-    if (config?.requires_discord_server_id) {
-      return contentIdInputConfig.labels.discordServerId;
-    }
-
-    if (config?.with_optional_chain_id) {
-      return contentIdInputConfig.labels.chainId;
-    }
-
-    if (config?.requires_group_id) {
-      return contentIdInputConfig.labels.groupId;
-    }
-
-    return 'Content Id';
-  };
-
-  const getContentIdInputPlaceholder = () => {
-    if (defaultValues?.contentIdScope === QuestActionContentIdScope.Thread) {
-      if (config?.with_optional_thread_id) {
-        return contentIdInputConfig.placeholders.sampleThreadLink;
-      }
-      if (config?.with_optional_comment_id) {
-        return contentIdInputConfig.placeholders.sampleCommentLink;
-      }
-    }
-    if (
-      config?.with_optional_topic_id ||
-      defaultValues?.contentIdScope === QuestActionContentIdScope.Topic
-    ) {
-      return contentIdInputConfig.placeholders.sampleTopicLink;
-    }
-
-    if (config?.requires_twitter_tweet_link) {
-      return contentIdInputConfig.placeholders.twitterTweetUrl;
-    }
-
-    if (config?.requires_discord_server_id) {
-      return contentIdInputConfig.placeholders.discordServerId;
-    }
-
-    if (config?.with_optional_chain_id) {
-      return contentIdInputConfig.placeholders.chainId;
-    }
-
-    if (config?.requires_group_id) {
-      return contentIdInputConfig.placeholders.groupId;
-    }
-
-    return 'Content Id';
-  };
-
-  const getStartLinkInputLabel = () => {
-    if (config?.requires_discord_server_id) {
-      return startLinkInputConfig.labels.discordServerUrl;
-    }
-
-    return 'Start Link';
-  };
-
-  const getStartLinkInputPlaceholder = () => {
-    if (config?.requires_discord_server_id) {
-      return startLinkInputConfig.placeholders.discordServerUrl;
-    }
-
-    return 'https://example.com';
-  };
-
-  const allowsContentId =
-    config?.with_optional_comment_id ||
-    config?.with_optional_thread_id ||
-    config?.with_optional_topic_id ||
-    config?.requires_twitter_tweet_link ||
-    config?.requires_discord_server_id ||
-    config?.with_optional_chain_id ||
-    config?.requires_group_id;
-
-  const repetitionCycleOptions = Object.keys(QuestParticipationPeriod).map(
-    (k) => ({
-      label: k,
-      value: QuestParticipationPeriod[k],
-    }),
-  );
-
-  const repetitionCycleValidatorFn = (props: ValidationFnProps) => {
-    const participation_limit = defaultValues?.participationLimit;
-    const { input, selectList } = props.values;
-
-    // clear errors if participation timeline is not a repeatable
-    if (participation_limit !== QuestParticipationLimit.OncePerPeriod) {
-      return { error: undefined };
-    }
-
-    // validate repetition cycle value
-    if (
-      !Object.values(QuestParticipationPeriod).includes(
-        selectList?.value as QuestParticipationPeriod,
-      )
-    ) {
-      return { error: 'Invalid value for reptition cycle' };
-    }
-
-    // validate repetition count value
-    try {
-      numberNonDecimalGTZeroValidationSchema.parse(`${input}`);
-
-      const count = parseInt(`${input}`);
-
-      // verify repetition counts fall within a certain range
-      if (
-        (selectList?.value === QuestParticipationPeriod.Daily &&
-          count > MAX_REPETITION_COUNTS.PER_DAY) ||
-        (selectList?.value === QuestParticipationPeriod.Weekly &&
-          count > MAX_REPETITION_COUNTS.PER_WEEK) ||
-        (selectList?.value === QuestParticipationPeriod.Monthly &&
-          count > MAX_REPETITION_COUNTS.PER_MONTH)
-      ) {
-        const allowedCount =
-          selectList?.value === QuestParticipationPeriod.Daily
-            ? MAX_REPETITION_COUNTS.PER_DAY
-            : selectList?.value === QuestParticipationPeriod.Weekly
-              ? MAX_REPETITION_COUNTS.PER_WEEK
-              : MAX_REPETITION_COUNTS.PER_MONTH;
-        return {
-          error: `Cannot repeat more than ${allowedCount} times ${selectList?.value}`,
-        };
-      }
-    } catch {
-      return { error: 'Invalid value for repetition count' };
-    }
-
-    return { error: undefined };
-  };
-
-  const chainNodes = fetchCachedNodes()
-    ?.filter((node) => node?.ethChainId || node?.cosmosChainId)
-    ?.map((node) => ({
-      label: `${node.name} - (Chain Id = ${node?.ethChainId || node?.cosmosChainId})`,
-      value: node.id,
-    }))
-    ?.sort((a, b) => a.label.localeCompare(b.label));
+const QuestActionSubForm = (props: QuestActionSubFormProps) => {
+  const {
+    config,
+    defaultValues,
+    errors,
+    isRemoveable = true,
+    onChange,
+    onRemove,
+  } = props;
 
   const {
-    error: repetitionCycleRadioError,
-    triggerValidation: triggerRepetitionCycleRadioValidation,
-    ...repetitionCycleRadioProps
-  } = useCWRepetitionCycleRadioButton({
-    validatorFn: repetitionCycleValidatorFn,
-    repetitionCycleInputProps: {
-      value: 1,
-    },
-    repetitionCycleSelectListProps: {
-      options: repetitionCycleOptions,
-      selected: repetitionCycleOptions[0],
-    },
-  });
-
-  const repetitionCycleRadio = {
-    error: repetitionCycleRadioError,
-    triggerValidation: triggerRepetitionCycleRadioValidation,
-    props: {
-      repetitionCycleInputProps: {
-        ...repetitionCycleRadioProps.repetitionCycleInputProps,
-      },
-      repetitionCycleSelectListProps: {
-        ...repetitionCycleRadioProps.repetitionCycleSelectListProps,
-      },
-    },
-  };
-
-  if (internalRefs) {
-    internalRefs.runParticipationLimitValidator =
-      triggerRepetitionCycleRadioValidation;
-  }
-
-  useRunOnceOnCondition({
-    callback: () => {
-      if (
-        defaultValues?.participationTimesPerPeriod ||
-        defaultValues?.participationPeriod
-      ) {
-        defaultValues?.participationTimesPerPeriod &&
-          repetitionCycleRadioProps.repetitionCycleInputProps.onChange(
-            defaultValues?.participationTimesPerPeriod,
-          );
-        defaultValues?.participationPeriod &&
-          repetitionCycleRadioProps.repetitionCycleSelectListProps.onChange({
-            value: defaultValues?.participationPeriod,
-            label:
-              Object.entries(QuestParticipationPeriod).find(
-                ([_, v]) => v === defaultValues?.participationPeriod,
-              )?.[0] || '',
-          });
-      }
-    },
-    shouldRun: true,
-  });
-
-  const participationTimesPerPeriod =
-    repetitionCycleRadioProps.repetitionCycleInputProps.value;
-  useEffect(() => {
-    if (
-      participationTimesPerPeriod === defaultValues?.participationTimesPerPeriod
-    )
-      return;
-    onChange?.({
-      participationTimesPerPeriod: participationTimesPerPeriod,
-    });
-  }, [
-    participationTimesPerPeriod,
-    defaultValues?.participationTimesPerPeriod,
-    onChange,
-  ]);
-
-  const participationPeriod = repetitionCycleRadioProps
-    .repetitionCycleSelectListProps.selected?.value as QuestParticipationPeriod;
-  useEffect(() => {
-    if (participationPeriod === defaultValues?.participationPeriod) return;
-    onChange?.({
-      participationPeriod: participationPeriod,
-    });
-  }, [participationPeriod, defaultValues?.participationPeriod, onChange]);
-
-  const doesPreventRepetition =
-    typeof config?.is_action_repeatable !== 'undefined'
-      ? !config?.is_action_repeatable
-      : false;
+    doesActionPreventRepetition,
+    repetitionCycleRadio,
+    actionOptions,
+    inputConfigs,
+    hasContentIdField,
+    communityChainNodeSelectInputOptions,
+  } = useQuestActionSubForm(props);
 
   return (
     <div className={clsx('QuestActionSubForm', { isRemoveable })}>
@@ -364,10 +69,10 @@ const QuestActionSubForm = ({
                 participationLimit: QuestParticipationLimit.OncePerPeriod,
               })
             }
-            disabled={doesPreventRepetition}
+            disabled={doesActionPreventRepetition}
           />,
           `Selected action does not allow repetition`,
-          doesPreventRepetition,
+          doesActionPreventRepetition,
           'w-fit',
         )}
         <CWRadioButton
@@ -513,9 +218,9 @@ const QuestActionSubForm = ({
 
       {config?.requires_start_link && (
         <CWTextInput
-          label={getStartLinkInputLabel()}
+          label={inputConfigs.startLink.label}
           name="startLink"
-          placeholder={getStartLinkInputPlaceholder()}
+          placeholder={inputConfigs.startLink.placeholder}
           fullWidth
           {...(defaultValues?.startLink && {
             defaultValue: defaultValues?.startLink,
@@ -565,9 +270,11 @@ const QuestActionSubForm = ({
         </div>
       )}
 
-      <div className={clsx('grid-row', allowsContentId ? 'cols-2' : 'cols-1')}>
-        {allowsContentId &&
-          (config.with_optional_chain_id ? (
+      <div
+        className={clsx('grid-row', hasContentIdField ? 'cols-2' : 'cols-1')}
+      >
+        {hasContentIdField &&
+          (config?.with_optional_chain_id ? (
             <CWSelectList
               isClearable={true}
               backspaceRemovesValue
@@ -575,7 +282,7 @@ const QuestActionSubForm = ({
               name="contentIdentifier"
               label="Chain Node"
               placeholder="Select a chain node"
-              options={chainNodes}
+              options={communityChainNodeSelectInputOptions}
               onChange={(newValue) =>
                 onChange?.({ contentIdentifier: `${newValue?.value || ''}` })
               }
@@ -583,7 +290,7 @@ const QuestActionSubForm = ({
                 value: {
                   value: parseInt(`${defaultValues?.contentIdentifier}`),
                   label: `${
-                    chainNodes?.find(
+                    communityChainNodeSelectInputOptions?.find(
                       (x) =>
                         x.value ===
                         parseInt(`${defaultValues?.contentIdentifier}`),
@@ -597,8 +304,8 @@ const QuestActionSubForm = ({
             <CWTextInput
               key={`contentIdentifier-${defaultValues?.action}-${defaultValues?.contentIdScope}`}
               name="contentIdentifier"
-              label={getContentIdInputLabel()}
-              placeholder={getContentIdInputPlaceholder()}
+              label={inputConfigs.contentId.label}
+              placeholder={inputConfigs.contentId.placeholder}
               fullWidth
               {...(defaultValues?.contentIdentifier && {
                 defaultValue: defaultValues?.contentIdentifier,
